Update readme, use clangd for cu

+++ pmpp/mat_mul_tiling.h
@@ -1,6 +1,6 @@
 #ifndef MAT_MUL_TILING_KERNEL_H
 #define MAT_MUL_TILING_KERNEL_H
 
-void matMulTiling(float* A, float* B, float* C, int i, int j, int k);
+void matMulTiling(float* A, float* B, float* C, size_t i, size_t j, size_t k);
 
 #endif
