+++ pmpp/vec_add.h
@@ -1,4 +1,4 @@
 #ifndef PMPP_VEC_ADD
 #define PMPP_VEC_ADD
-void vecAdd(float* A, float* B, float* C, int n);
+void vecAdd(float* A, float* B, float* C, size_t n);
 #endif
